requirements.txt
accelerate
blobfile
datasets
httpx[http2]
mcp[cli]
memray  # needed for debugging (but is lightweight), we can put it to dev mode when using pyproject.toml
omegaconf
pillow
pybase64
pylatexenc
pytest-asyncio
pyyaml
qwen_vl_utils # for VLM
ray[default]
ring_flash_attn
sglang-router>=0.2.3
tensorboard
torchft-nightly==2026.4.3
transformers
wandb
